Clarify the output
The syncer prints the example repository path that is being processed. When
done processing, it will display a summary of the changes made to the master
and gh-pages branches.

When all the example application have been processed, the syncer displays a
summary of the number of repositories that have been updated.